restructured to split code and headers, made cmake file slightly more generic, added #def for prettier debug printing
