+++ cpp/.gitignore
@@ -1,3 +1,7 @@
 cubes
-*.bin
+cubes_*.bin
 build/**
+CMakeFiles
+cmake_install.cmake
+CMakeCache.txt
+Makefile
