+++ cpp/CMakeLists.txt
@@ -4,7 +4,10 @@ project(cubes)
 set(CMAKE_BUILD_TYPE "Release")
 set(CMAKE_CXX_FLAGS "-std=c++17")
 set(CMAKE_CXX_FLAGS_RELEASE "-march=native -O3 -Wall -Wextra")
+set(CMAKE_CXX_FLAGS_DEBUG "-march=native -O -Wall -Werror -Wextra -D DEBUG")
 
-add_executable(${PROJECT_NAME} cubes.cpp)
+include_directories("include")
+file(GLOB SOURCES "src/*.cpp")
+add_executable(${PROJECT_NAME} ${SOURCES})
 target_link_libraries(${PROJECT_NAME} pthread)
 target_compile_definitions(${PROJECT_NAME} PUBLIC NDEBUG)
